requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
openpyxl>=3.1.2
pydantic>=2.5.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
langdetect>=1.0.9
deep-translator>=1.11.4

# FX rates
forex-python>=1.8

# LLM for bootstrapper (optional, user can choose)
anthropic>=0.18.0
openai>=1.12.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
freezegun>=1.4.0

# Utilities
tenacity>=8.2.3
ratelimit>=2.2.1
urllib3>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
